requirements.txt
Django==2.2
django-countries==5.5
django-debug-toolbar==2.2
djangorestframework==3.11.0
Pillow==7.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
django-rest-swagger==2.2.0
psycopg2-binary==2.8.3
django-extensions==2.2.5